+++ tests/conftest.py
@@ -382,3 +382,18 @@ def desired_delivery_date():
 @pytest.fixture
 def billing():
     return read_fixture_data()["billing"]
+
+
+@pytest.fixture
+def luma_ruleset_name():
+    return "required_deliver_by_date"
+
+
+@pytest.fixture
+def luma_planned_ship_date():
+    return "2025-06-12"
+
+
+@pytest.fixture
+def luma_deliver_by_date():
+    return "2025-06-19"